Move game_def to common dir
