+++ src/common/game_def.hpp
